feat(Table): now supports pinning
header and/or column
