+++ demo/app/components/table/table.component.scss
@@ -1,0 +1,4 @@
+.example-container {
+  height: 400px;
+  overflow: auto;
+}
