+++ terminus-ui/table/src/table.component.html
@@ -1,7 +1,3 @@
-<div class="ts-table__container">
-  <div class="ts-table__table">
-    <ng-container headerRowOutlet></ng-container>
-    <ng-container rowOutlet></ng-container>
-    <ng-container footerRowOutlet></ng-container>
-  </div>
-</div>
+<ng-container headerRowOutlet></ng-container>
+<ng-container rowOutlet></ng-container>
+<ng-container footerRowOutlet></ng-container>
