+++ terminus-ui/table/src/table.component.spec.ts
@@ -1,349 +1,18 @@
 /* eslint-disable no-underscore-dangle */
-import { DataSource } from '@angular/cdk/collections';
-import {
-  Component,
-  Provider,
-  Type,
-  ViewChild,
-} from '@angular/core';
 import { ComponentFixture } from '@angular/core/testing';
+import { createComponent } from '@terminus/ngx-tools/testing';
+import * as testComponents from '@terminus/ui/table/testing';
+// eslint-disable-next-line no-duplicate-imports
 import {
-  FormsModule,
-  ReactiveFormsModule,
-} from '@angular/forms';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { TsWindowService } from '@terminus/ngx-tools/browser';
-import { TsWindowServiceMock } from '@terminus/ngx-tools/browser/testing';
-import { createComponent as createComponentInner } from '@terminus/ngx-tools/testing';
-import {
-  TsPaginatorComponent,
-  TsPaginatorModule,
-} from '@terminus/ui/paginator';
-import {
-  TsSortDirective,
-  TsSortHeaderComponent,
-  TsSortModule,
-} from '@terminus/ui/sort';
-import {
-  BehaviorSubject,
-  Observable,
-} from 'rxjs';
+  expectTableToMatchContent,
+  getHeaderCells,
+  getHeaderRow,
+} from '@terminus/ui/table/testing';
+
 import { TsTableDataSource } from './table-data-source';
-import { TsTableComponent } from './table.component';
 import { TsTableModule } from './table.module';
 
 
-
-
-/**
- * HELPERS
- */
-
-
-interface TestData {
-  a: string;
-  b: string;
-  c: string;
-}
-
-// TODO: change to my datasource - says properties connect aren't the same???
-class FakeDataSource extends DataSource<TestData> {
-  public _dataChange = new BehaviorSubject<TestData[]>([]);
-  public set data(data: TestData[]) {
-    this._dataChange.next(data);
-  }
-  public get data() {
-    return this._dataChange.getValue();
-  }
-
-  public constructor() {
-    super();
-    for (let i = 0; i < 4; i++) {
-      this.addData();
-    }
-  }
-
-  public connect(): Observable<TestData[]> {
-    return this._dataChange;
-  }
-
-  public disconnect() {}
-
-  public addData() {
-    const nextIndex = this.data.length + 1;
-
-    const copiedData = this.data.slice();
-    copiedData.push({
-      a: `a_${nextIndex}`,
-      b: `b_${nextIndex}`,
-      c: `c_${nextIndex}`,
-    });
-
-    this.data = copiedData;
-  }
-}
-
-@Component({
-  template: `
-    <ts-table [dataSource]="dataSource">
-      <ng-container tsColumnDef="column_a">
-        <ts-header-cell *tsHeaderCellDef> Column A</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.a }}</ts-cell>
-      </ng-container>
-
-      <ng-container tsColumnDef="column_b">
-        <ts-header-cell *tsHeaderCellDef> Column B</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.b }}</ts-cell>
-      </ng-container>
-
-      <ng-container tsColumnDef="column_c">
-        <ts-header-cell *tsHeaderCellDef> Column C</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.c }}</ts-cell>
-      </ng-container>
-
-      <ng-container tsColumnDef="special_column">
-        <ts-cell *tsCellDef="let row">fourth_row</ts-cell>
-      </ng-container>
-
-      <ts-header-row *tsHeaderRowDef="columnsToRender"></ts-header-row>
-      <ts-row *tsRowDef="let row; columns: columnsToRender"></ts-row>
-      <ts-row *tsRowDef="let row; columns: ['special_column']; when: isFourthRow"></ts-row>
-    </ts-table>
-  `,
-})
-class TableApp {
-  @ViewChild(TsTableComponent, {static: true})
-  public table!: TsTableComponent<TestData>;
-
-  public dataSource: FakeDataSource | null = new FakeDataSource();
-  public columnsToRender = ['column_a', 'column_b', 'column_c'];
-  public isFourthRow = (i: number, _rowData: TestData) => i === 3;
-
-}
-
-
-@Component({
-  template: `
-    <ts-table [dataSource]="dataSource">
-      <ng-container tsColumnDef="column_a">
-        <ts-header-cell *tsHeaderCellDef> Column A</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.a }}</ts-cell>
-      </ng-container>
-
-      <ng-container tsColumnDef="special_column">
-        <ts-cell *tsCellDef="let row">fourth_row</ts-cell>
-      </ng-container>
-
-      <ts-header-row *tsHeaderRowDef="['column_a']"></ts-header-row>
-      <ts-row *tsRowDef="let row; columns: ['column_a']"></ts-row>
-      <ts-row *tsRowDef="let row; columns: ['special_column']; when: isFourthRow"></ts-row>
-    </ts-table>
-  `,
-})
-class TableWithWhenRowApp {
-  @ViewChild(TsTableComponent, {static: true})
-  public table!: TsTableComponent<TestData>;
-  public dataSource: FakeDataSource | null = new FakeDataSource();
-  public isFourthRow = (i: number, _rowData: TestData) => i === 3;
-}
-
-
-@Component({
-  template: `
-    <ts-table [dataSource]="dataSource" tsSort>
-      <ng-container tsColumnDef="column_a" noWrap="true">
-        <ts-header-cell *tsHeaderCellDef ts-sort-header="a">Column A</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.a }}</ts-cell>
-      </ng-container>
-
-      <ng-container tsColumnDef="column_b" minWidth="100px"  noWrap="true">
-        <ts-header-cell *tsHeaderCellDef>Column B</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.b }}</ts-cell>
-      </ng-container>
-
-      <ng-container tsColumnDef="column_c">
-        <ts-header-cell *tsHeaderCellDef>Column C</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.c }}</ts-cell>
-      </ng-container>
-
-      <ts-header-row *tsHeaderRowDef="columnsToRender"></ts-header-row>
-      <ts-row *tsRowDef="let row; columns: columnsToRender"></ts-row>
-    </ts-table>
-
-    <ts-paginator></ts-paginator>
-  `,
-})
-class ArrayDataSourceTableApp {
-  public underlyingDataSource = new FakeDataSource();
-  public dataSource = new TsTableDataSource<TestData>();
-  public columnsToRender = ['column_a', 'column_b', 'column_c'];
-
-  @ViewChild(TsTableComponent, {static: true})
-  public table!: TsTableComponent<TestData>;
-  @ViewChild(TsPaginatorComponent, {static: true})
-  public paginator!: TsPaginatorComponent;
-  @ViewChild(TsSortDirective, {static: true})
-  public sort!: TsSortDirective;
-  @ViewChild(TsSortHeaderComponent, {static: true})
-  public sortHeader!: TsSortHeaderComponent;
-
-  public constructor() {
-    this.underlyingDataSource.data = [];
-
-    // Add three rows of data
-    this.underlyingDataSource.addData();
-    this.underlyingDataSource.addData();
-    this.underlyingDataSource.addData();
-
-    this.underlyingDataSource.connect().subscribe(data => {
-      this.dataSource.data = data;
-    });
-  }
-
-}
-
-@Component({
-  template: `
-    <ts-table [dataSource]="dataSource" tsSort>
-      <ng-container tsColumnDef="column_a" alignment="left">
-        <ts-header-cell *tsHeaderCellDef>Column A</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.a }}</ts-cell>
-      </ng-container>
-
-      <ng-container tsColumnDef="column_b">
-        <ts-header-cell *tsHeaderCellDef>Column B</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.b }}</ts-cell>
-      </ng-container>
-
-      <ng-container tsColumnDef="column_c">
-        <ts-header-cell *tsHeaderCellDef>Column C</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.c }}</ts-cell>
-      </ng-container>
-
-      <ts-header-row *tsHeaderRowDef="columnsToRender"></ts-header-row>
-      <ts-row *tsRowDef="let row; columns: columnsToRender"></ts-row>
-    </ts-table>
-  `,
-})
-class TableColumnAlignmentTableApp {
-  public underlyingDataSource = new FakeDataSource();
-  public dataSource = new TsTableDataSource<TestData>();
-  public columnsToRender = ['column_a', 'column_b', 'column_c'];
-
-  @ViewChild(TsTableComponent, {static: true})
-  public table!: TsTableComponent<TestData>;
-
-  public constructor() {
-    this.underlyingDataSource.data = [];
-
-    // Add a row of data
-    this.underlyingDataSource.addData();
-
-    this.underlyingDataSource.connect().subscribe(data => {
-      this.dataSource.data = data;
-    });
-  }
-
-}
-
-@Component({
-  template: `
-    <ts-table [dataSource]="dataSource" tsSort>
-      <ng-container tsColumnDef="column_a" alignment="top">
-        <ts-header-cell *tsHeaderCellDef>Column A</ts-header-cell>
-        <ts-cell *tsCellDef="let row">{{ row.a }}</ts-cell>
-      </ng-container>
-      <ts-header-row *tsHeaderRowDef="columnsToRender"></ts-header-row>
-      <ts-row *tsRowDef="let row; columns: columnsToRender"></ts-row>
-    </ts-table>
-  `,
-})
-class TableColumnInvalidAlignmentTableApp {
-  public underlyingDataSource = new FakeDataSource();
-  public dataSource = new TsTableDataSource<TestData>();
-  public columnsToRender = ['column_a'];
-
-  @ViewChild(TsTableComponent, {static: true})
-  public table!: TsTableComponent<TestData>;
-
-  public constructor() {
-    this.underlyingDataSource.data = [];
-
-    // Add a row of data
-    this.underlyingDataSource.addData();
-
-    this.underlyingDataSource.connect().subscribe(data => {
-      this.dataSource.data = data;
-    });
-  }
-
-}
-
-
-// Utilities copied from CDKTable's spec
-function getElements(element: Element, query: string): Element[] {
-  return [].slice.call(element.querySelectorAll(query));
-}
-
-function getHeaderRow(tableElement: Element): Element {
-  return tableElement.querySelector('.ts-header-row')!;
-}
-
-function getRows(tableElement: Element): Element[] {
-  return getElements(tableElement, '.ts-row');
-}
-function getCells(row: Element): Element[] {
-  return row ? getElements(row, '.ts-cell') : [];
-}
-
-function getHeaderCells(tableElement: Element): Element[] {
-  return getElements(getHeaderRow(tableElement), '.ts-header-cell');
-}
-
-function expectTableToMatchContent(tableElement: Element, expectedTableContent: any[]) {
-  const missedExpectations: string[] = [];
-  function checkCellContent(cell: Element, expectedTextContent: string) {
-    const actualTextContent = cell.textContent!.trim();
-    if (actualTextContent !== expectedTextContent) {
-      missedExpectations.push(
-        `Expected cell contents to be ${expectedTextContent} but was ${actualTextContent}`
-      );
-    }
-  }
-
-  // Check header cells
-  const expectedHeaderContent = expectedTableContent.shift();
-  getHeaderCells(tableElement).forEach((cell, index) => {
-    const expected = expectedHeaderContent
-      ? expectedHeaderContent[index]
-      : null;
-    checkCellContent(cell, expected);
-  });
-
-  // Check data row cells
-  const rows = getRows(tableElement);
-  expect(rows.length).toBe(expectedTableContent.length);
-  rows.forEach((row, rowIndex) => {
-    getCells(row).forEach((cell, cellIndex) => {
-      const expected = expectedTableContent.length
-        ? expectedTableContent[rowIndex][cellIndex]
-        : null;
-      checkCellContent(cell, expected);
-    });
-  });
-
-  if (missedExpectations.length) {
-    // eslint-disable-next-line no-undef
-    fail(missedExpectations.join('\n'));
-  }
-}
-
-
-
-
-/**
- * TESTS
- */
 
 
 describe(`TsTableComponent`, function() {
@@ -351,11 +20,11 @@ describe(`TsTableComponent`, function() {
   describe(`with basic data source`, function() {
 
     test(`should be able to create a table with the right content and without when row`, function() {
-      const fixture = createComponent(TableApp);
+      const fixture = createComponent(testComponents.TableApp, [], [TsTableModule]);
       fixture.detectChanges();
 
       const tableElement = fixture.nativeElement.querySelector('.ts-table');
-      const data = fixture.componentInstance.dataSource!.data;
+      const data = fixture.componentInstance.dataSource.data;
       expectTableToMatchContent(tableElement, [
         ['Column A', 'Column B', 'Column C'],
         [data[0].a, data[0].b, data[0].c],
@@ -367,7 +36,7 @@ describe(`TsTableComponent`, function() {
 
 
     test(`should create a table with special when row`, function() {
-      const fixture = createComponent(TableWithWhenRowApp);
+      const fixture = createComponent(testComponents.TableWithWhenRowApp, [], [TsTableModule]);
       fixture.detectChanges();
 
       const tableElement = document.querySelector('.ts-table');
@@ -384,12 +53,12 @@ describe(`TsTableComponent`, function() {
 
   describe(`with TsTableDataSource`, function() {
     let tableElement: HTMLElement;
-    let fixture: ComponentFixture<ArrayDataSourceTableApp>;
+    let fixture: ComponentFixture<testComponents.ArrayDataSourceTableApp>;
     let dataSource: TsTableDataSource<TestData>;
-    let component: ArrayDataSourceTableApp;
+    let component: testComponents.ArrayDataSourceTableApp;
 
     beforeEach(() => {
-      fixture = createComponent(ArrayDataSourceTableApp);
+      fixture = createComponent(testComponents.ArrayDataSourceTableApp, [], [TsTableModule]);
       fixture.detectChanges();
 
       tableElement = fixture.nativeElement.querySelector('.ts-table');
@@ -443,22 +112,27 @@ describe(`TsTableComponent`, function() {
 
     test(`should add the min-width style`, function() {
       const column = fixture.nativeElement.querySelector('.ts-cell.ts-column-column_b');
+      const headerColumn = fixture.nativeElement.querySelector('.ts-header-cell.ts-column-column_b');
 
       let style;
+      let headerStyle;
       if (column.style && column.style._values) {
-        style = column.style._values['min-width'];
+        style = column.style._values['flex-basis'];
+        headerStyle = headerColumn.style._values['flex-basis'];
       }
 
       expect(style).toEqual('100px');
+      expect(headerStyle).toEqual('100px');
     });
 
   });
 
+
   describe(`table column alignment`, () => {
-    let fixture: ComponentFixture<TableColumnAlignmentTableApp>;
+    let fixture: ComponentFixture<testComponents.TableColumnAlignmentTableApp>;
 
     beforeEach(() => {
-      fixture = createComponent(TableColumnAlignmentTableApp);
+      fixture = createComponent(testComponents.TableColumnAlignmentTableApp, [], [TsTableModule]);
       fixture.detectChanges();
     });
 
@@ -502,7 +176,7 @@ describe(`TsTableComponent`, function() {
 
     test(`should throw warning`, () => {
       window.console.warn = jest.fn();
-      const fixture = createComponent(TableColumnInvalidAlignmentTableApp);
+      const fixture = createComponent(testComponents.TableColumnInvalidAlignmentTableApp, [], [TsTableModule]);
       fixture.detectChanges();
 
       expect(window.console.warn).toHaveBeenCalled();
@@ -510,31 +184,30 @@ describe(`TsTableComponent`, function() {
 
   });
 
+
+  describe(`pinned header and column`, () => {
+    let fixture: ComponentFixture<testComponents.PinnedTableHeaderColumn>;
+    let tableElement: HTMLElement;
+
+    beforeEach(() => {
+      fixture = createComponent(testComponents.PinnedTableHeaderColumn, [], [TsTableModule]);
+      fixture.detectChanges();
+
+      tableElement = fixture.nativeElement.querySelector('.ts-table');
+    });
+
+    test(`should set a header to be sticky`, () => {
+      const header = getHeaderRow(tableElement);
+      expect(header.classList).toContain('ts-table--sticky');
+    });
+
+    test(`should set a column to be sticky`, () => {
+      const headerCells = getHeaderCells(tableElement);
+      expect(headerCells[0].classList).toContain('ts-table--sticky');
+      expect(headerCells[1].classList).not.toContain('ts-table--sticky');
+    });
+
+  });
+
 });
 
-
-
-
-/**
- * HELPERS
- */
-
-export function createComponent<T>(component: Type<T>, providers: Provider[] = [], imports: any[] = []): ComponentFixture<T> {
-  return createComponentInner<T>(component,
-    [
-      {
-        provide: TsWindowService,
-        useClass: TsWindowServiceMock,
-        ...providers,
-      },
-    ],
-    [
-      NoopAnimationsModule,
-      FormsModule,
-      ReactiveFormsModule,
-      TsTableModule,
-      TsPaginatorModule,
-      TsSortModule,
-      ...imports,
-    ],);
-}
