+++ terminus-ui/table/testing/src/public-api.ts
@@ -1,0 +1,2 @@
+export * from './test-components';
+export * from './test-helpers';
