+++ terminus-ui/tsconfig.spec.json
@@ -11,43 +11,45 @@
       "jest"
     ],
     "paths": {
-      "@terminus/ui/autocomplete"        : ["terminus-ui/autocomplete/src/public-api.ts"],
-      "@terminus/ui/autofocus"           : ["terminus-ui/autofocus/src/public-api.ts"],
-      "@terminus/ui/button"              : ["terminus-ui/button/src/public-api.ts"],
-      "@terminus/ui/card"                : ["terminus-ui/card/src/public-api.ts"],
-      "@terminus/ui/chart"               : ["terminus-ui/chart/src/public-api.ts"],
-      "@terminus/ui/checkbox"            : ["terminus-ui/checkbox/src/public-api.ts"],
-      "@terminus/ui/confirmation"        : ["terminus-ui/confirmation/src/public-api.ts"],
-      "@terminus/ui/copy"                : ["terminus-ui/copy/src/public-api.ts"],
-      "@terminus/ui/csv-entry"           : ["terminus-ui/csv-entry/src/public-api.ts"],
-      "@terminus/ui/date-range"          : ["terminus-ui/date-range/src/public-api.ts"],
-      "@terminus/ui/date-range/testing"  : ["terminus-ui/date-range/testing/src/public-api.ts"],
-      "@terminus/ui/file-upload"         : ["terminus-ui/file-upload/src/public-api.ts"],
-      "@terminus/ui/form-field"          : ["terminus-ui/form-field/src/public-api.ts"],
-      "@terminus/ui/icon-button"         : ["terminus-ui/icon-button/src/public-api.ts"],
-      "@terminus/ui/icon"                : ["terminus-ui/icon/src/public-api.ts"],
-      "@terminus/ui/input"               : ["terminus-ui/input/src/public-api.ts"],
-      "@terminus/ui/input/testing"       : ["terminus-ui/input/testing/src/public-api.ts"],
-      "@terminus/ui/link"                : ["terminus-ui/link/src/public-api.ts"],
-      "@terminus/ui/loading-overlay"     : ["terminus-ui/loading-overlay/src/public-api.ts"],
-      "@terminus/ui/login-form"          : ["terminus-ui/login-form/src/public-api.ts"],
-      "@terminus/ui/menu"                : ["terminus-ui/menu/src/public-api.ts"],
-      "@terminus/ui/navigation"          : ["terminus-ui/navigation/src/public-api.ts"],
-      "@terminus/ui/paginator"           : ["terminus-ui/paginator/src/public-api.ts"],
-      "@terminus/ui/pipes"               : ["terminus-ui/pipes/src/public-api.ts"],
-      "@terminus/ui/radio-group"         : ["terminus-ui/radio-group/src/public-api.ts"],
-      "@terminus/ui/scrollbars"          : ["terminus-ui/scrollbars/src/public-api.ts"],
-      "@terminus/ui/search"              : ["terminus-ui/search/src/public-api.ts"],
-      "@terminus/ui/select"              : ["terminus-ui/select/src/public-api.ts"],
-      "@terminus/ui/select/testing"      : ["terminus-ui/select/testing/src/public-api.ts"],
-      "@terminus/ui/sort"                : ["terminus-ui/sort/src/public-api.ts"],
-      "@terminus/ui/spacing"             : ["terminus-ui/spacing/src/public-api.ts"],
-      "@terminus/ui/table"               : ["terminus-ui/table/src/public-api.ts"],
-      "@terminus/ui/toggle"              : ["terminus-ui/toggle/src/public-api.ts"],
-      "@terminus/ui/tooltip"             : ["terminus-ui/tooltip/src/public-api.ts"],
-      "@terminus/ui/utilities"           : ["terminus-ui/utilities/src/public-api.ts"],
-      "@terminus/ui/validation-messages" : ["terminus-ui/validation-messages/src/public-api.ts"],
-      "@terminus/ui/validators"          : ["terminus-ui/validators/src/public-api.ts"]
+      "@terminus/ui/autocomplete"         : ["terminus-ui/autocomplete/src/public-api.ts"],
+      "@terminus/ui/autocomplete/testing" : ["terminus-ui/autocomplete/testing/src/public-api.ts"],
+      "@terminus/ui/autofocus"            : ["terminus-ui/autofocus/src/public-api.ts"],
+      "@terminus/ui/button"               : ["terminus-ui/button/src/public-api.ts"],
+      "@terminus/ui/card"                 : ["terminus-ui/card/src/public-api.ts"],
+      "@terminus/ui/chart"                : ["terminus-ui/chart/src/public-api.ts"],
+      "@terminus/ui/checkbox"             : ["terminus-ui/checkbox/src/public-api.ts"],
+      "@terminus/ui/confirmation"         : ["terminus-ui/confirmation/src/public-api.ts"],
+      "@terminus/ui/copy"                 : ["terminus-ui/copy/src/public-api.ts"],
+      "@terminus/ui/csv-entry"            : ["terminus-ui/csv-entry/src/public-api.ts"],
+      "@terminus/ui/date-range"           : ["terminus-ui/date-range/src/public-api.ts"],
+      "@terminus/ui/date-range/testing"   : ["terminus-ui/date-range/testing/src/public-api.ts"],
+      "@terminus/ui/file-upload"          : ["terminus-ui/file-upload/src/public-api.ts"],
+      "@terminus/ui/form-field"           : ["terminus-ui/form-field/src/public-api.ts"],
+      "@terminus/ui/icon-button"          : ["terminus-ui/icon-button/src/public-api.ts"],
+      "@terminus/ui/icon"                 : ["terminus-ui/icon/src/public-api.ts"],
+      "@terminus/ui/input"                : ["terminus-ui/input/src/public-api.ts"],
+      "@terminus/ui/input/testing"        : ["terminus-ui/input/testing/src/public-api.ts"],
+      "@terminus/ui/link"                 : ["terminus-ui/link/src/public-api.ts"],
+      "@terminus/ui/loading-overlay"      : ["terminus-ui/loading-overlay/src/public-api.ts"],
+      "@terminus/ui/login-form"           : ["terminus-ui/login-form/src/public-api.ts"],
+      "@terminus/ui/menu"                 : ["terminus-ui/menu/src/public-api.ts"],
+      "@terminus/ui/navigation"           : ["terminus-ui/navigation/src/public-api.ts"],
+      "@terminus/ui/paginator"            : ["terminus-ui/paginator/src/public-api.ts"],
+      "@terminus/ui/pipes"                : ["terminus-ui/pipes/src/public-api.ts"],
+      "@terminus/ui/radio-group"          : ["terminus-ui/radio-group/src/public-api.ts"],
+      "@terminus/ui/scrollbars"           : ["terminus-ui/scrollbars/src/public-api.ts"],
+      "@terminus/ui/search"               : ["terminus-ui/search/src/public-api.ts"],
+      "@terminus/ui/select"               : ["terminus-ui/select/src/public-api.ts"],
+      "@terminus/ui/select/testing"       : ["terminus-ui/select/testing/src/public-api.ts"],
+      "@terminus/ui/sort"                 : ["terminus-ui/sort/src/public-api.ts"],
+      "@terminus/ui/spacing"              : ["terminus-ui/spacing/src/public-api.ts"],
+      "@terminus/ui/table"                : ["terminus-ui/table/src/public-api.ts"],
+      "@terminus/ui/table/testing"        : ["terminus-ui/table/testing/src/public-api.ts"],
+      "@terminus/ui/toggle"               : ["terminus-ui/toggle/src/public-api.ts"],
+      "@terminus/ui/tooltip"              : ["terminus-ui/tooltip/src/public-api.ts"],
+      "@terminus/ui/utilities"            : ["terminus-ui/utilities/src/public-api.ts"],
+      "@terminus/ui/validation-messages"  : ["terminus-ui/validation-messages/src/public-api.ts"],
+      "@terminus/ui/validators"           : ["terminus-ui/validators/src/public-api.ts"]
     }
   }
 }
